+++ packages/frontend/src/components/NavBar/AiAgentsButton.tsx
@@ -1,18 +1,45 @@
 import { CommercialFeatureFlags } from '@lightdash/common';
-import { Button } from '@mantine/core';
+import { Button as MantineButton } from '@mantine/core';
 import { IconMessageCircleStar } from '@tabler/icons-react';
 import { useNavigate } from 'react-router';
 import { useAiAgentPermission } from '../../ee/features/aiCopilot/hooks/useAiAgentPermission';
+import { useDefaultAgent } from '../../ee/features/aiCopilot/hooks/useDefaultAgent';
 import { useActiveProject } from '../../hooks/useActiveProject';
 import { useFeatureFlag } from '../../hooks/useFeatureFlagEnabled';
 import useApp from '../../providers/App/useApp';
 import MantineIcon from '../common/MantineIcon';
 
-export const AiAgentsButton = () => {
+const Button = ({ projectUuid }: { projectUuid: string }) => {
     // Using `navigate` instead of the `Link` component to ensure round corners within a button group
     const navigate = useNavigate();
+    const { defaultAgentUuid } = useDefaultAgent(projectUuid);
 
-    const { data: project } = useActiveProject();
+    const handleOnClick = () => {
+        if (defaultAgentUuid) {
+            void navigate(
+                `/projects/${projectUuid}/ai-agents/${defaultAgentUuid}`,
+            );
+        } else {
+            void navigate(`/projects/${projectUuid}/ai-agents`);
+        }
+    };
+    return (
+        <MantineButton
+            size="xs"
+            variant="default"
+            fz="sm"
+            leftIcon={
+                <MantineIcon icon={IconMessageCircleStar} color="#adb5bd" />
+            }
+            onClick={handleOnClick}
+        >
+            Ask AI
+        </MantineButton>
+    );
+};
+
+export const AiAgentsButton = () => {
+    const { data: projectUuid } = useActiveProject();
     const canViewAiAgents = useAiAgentPermission({ action: 'view' });
     const appQuery = useApp();
     const aiCopilotFlagQuery = useFeatureFlag(CommercialFeatureFlags.AiCopilot);
@@ -33,22 +60,10 @@ export const AiAgentsButton = () => {
         !canViewAiAgents ||
         !isAiCopilotEnabled ||
         !isAiAgentEnabled ||
-        !project
+        !projectUuid
     ) {
         return null;
     }
 
-    return (
-        <Button
-            size="xs"
-            variant="default"
-            fz="sm"
-            leftIcon={
-                <MantineIcon icon={IconMessageCircleStar} color="#adb5bd" />
-            }
-            onClick={() => navigate(`/projects/${project}/ai-agents`)}
-        >
-            Ask AI
-        </Button>
-    );
+    return <Button projectUuid={projectUuid} />;
 };