+++ packages/frontend/src/ee/features/aiCopilot/components/ChatElements/AgentChatBubbles.tsx
@@ -1,7 +1,9 @@
 import {
+    AiChartType,
     type AiAgentMessageAssistant,
     type AiAgentMessageUser,
     type AiAgentUser,
+    type FilterSchemaType,
 } from '@lightdash/common';
 import {
     ActionIcon,
@@ -31,13 +33,17 @@ import dayjs from 'dayjs';
 import { memo, useCallback, useMemo, type FC } from 'react';
 import { Link, useParams } from 'react-router';
 import MantineIcon from '../../../../../components/common/MantineIcon';
-import { useActiveProjectUuid } from '../../../../../hooks/useActiveProject';
+import ErrorBoundary from '../../../../../features/errorBoundary/ErrorBoundary';
+import { useInfiniteQueryResults } from '../../../../../hooks/useQueryResults';
 import { useTimeAgo } from '../../../../../hooks/useTimeAgo';
 import useApp from '../../../../../providers/App/useApp';
 import {
-    useAiAgentThreadMessageViz,
+    useAiAgentThreadMessageVizQuery,
     useUpdatePromptFeedbackMutation,
 } from '../../hooks/useOrganizationAiAgents';
+import { getOpenInExploreUrl } from '../../utils/getOpenInExploreUrl';
+import AgentVisualizationFilters from './AgentVisualizationFilters';
+import AgentVisualizationMetricsAndDimensions from './AgentVisualizationMetricsAndDimensions';
 import { AiChartVisualization } from './AiChartVisualization';
 
 export const UserBubble: FC<{ message: AiAgentMessageUser<AiAgentUser> }> = ({
@@ -88,14 +94,22 @@ export const AssistantBubble: FC<{
     message: AiAgentMessageAssistant;
     isPreview?: boolean;
 }> = memo(({ message, isPreview = false }) => {
-    const { agentUuid } = useParams();
-    const { activeProjectUuid } = useActiveProjectUuid();
+    const { agentUuid, projectUuid } = useParams();
 
-    const vizQuery = useAiAgentThreadMessageViz({
+    const queryExecutionHandle = useAiAgentThreadMessageVizQuery({
         agentUuid: agentUuid!,
         message,
-        activeProjectUuid,
+        projectUuid,
     });
+
+    const queryResults = useInfiniteQueryResults(
+        projectUuid,
+        queryExecutionHandle?.data?.query.queryUuid,
+    );
+
+    const isQueryLoading =
+        queryExecutionHandle.isLoading || queryResults.isFetchingRows;
+    const isQueryError = queryExecutionHandle.isError || queryResults.error;
 
     const updateFeedbackMutation = useUpdatePromptFeedbackMutation(
         agentUuid,
@@ -114,35 +128,55 @@ export const AssistantBubble: FC<{
     );
 
     const handleUpvote = useCallback(() => {
-        if (!activeProjectUuid || message.humanScore !== null) return; // Prevent changes if already rated
+        if (!projectUuid || message.humanScore !== null) return; // Prevent changes if already rated
         updateFeedbackMutation.mutate({
             messageUuid: message.uuid,
             humanScore: 1,
         });
-    }, [
-        activeProjectUuid,
-        message.uuid,
-        message.humanScore,
-        updateFeedbackMutation,
-    ]);
+    }, [projectUuid, message.uuid, message.humanScore, updateFeedbackMutation]);
 
     const handleDownvote = useCallback(() => {
-        if (!activeProjectUuid || message.humanScore !== null) return; // Prevent changes if already rated
+        if (!projectUuid || message.humanScore !== null) return; // Prevent changes if already rated
         updateFeedbackMutation.mutate({
             messageUuid: message.uuid,
             humanScore: -1,
         });
-    }, [
-        activeProjectUuid,
-        message.uuid,
-        message.humanScore,
-        updateFeedbackMutation,
-    ]);
+    }, [projectUuid, message.uuid, message.humanScore, updateFeedbackMutation]);
 
     // TODO: Do not use hardcoded string for loading state
     const isLoading = message.message === 'Thinking...';
     const hasRating =
         message.humanScore !== undefined && message.humanScore !== null;
+
+    const openInExploreUrl = useMemo(() => {
+        if (isQueryLoading || isQueryError) return '';
+
+        return getOpenInExploreUrl({
+            metricQuery: queryExecutionHandle.data.query.metricQuery,
+            projectUuid,
+            columnOrder: [
+                ...queryExecutionHandle.data.query.metricQuery.dimensions,
+                ...queryExecutionHandle.data.query.metricQuery.metrics,
+            ],
+            type: queryExecutionHandle.data.type,
+            vizConfig: message.vizConfigOutput,
+            rows: queryResults.rows,
+            pivotColumns:
+                // TODO: fix this using schema
+                message.vizConfigOutput &&
+                'breakdownByDimension' in message.vizConfigOutput &&
+                typeof message.vizConfigOutput.breakdownByDimension === 'string'
+                    ? [message.vizConfigOutput.breakdownByDimension]
+                    : undefined,
+        });
+    }, [
+        isQueryLoading,
+        isQueryError,
+        projectUuid,
+        queryExecutionHandle.data,
+        queryResults.rows,
+        message.vizConfigOutput,
+    ]);
 
     return (
         <Stack
@@ -173,18 +207,19 @@ export const AssistantBubble: FC<{
                     radius="md"
                     p="md"
                     shadow="none"
-                    {...((vizQuery.isError || vizQuery.isLoading) && {
+                    {...((queryExecutionHandle.isError ||
+                        queryExecutionHandle.isLoading) && {
                         bg: 'gray.0',
                         style: {
                             borderStyle: 'dashed',
                         },
                     })}
                 >
-                    {vizQuery.isLoading ? (
+                    {isQueryLoading ? (
                         <Center>
                             <Loader type="dots" color="gray" />
                         </Center>
-                    ) : vizQuery.isError ? (
+                    ) : isQueryError ? (
                         <Stack gap="xs" align="center">
                             <MantineIcon
                                 icon={IconExclamationCircle}
@@ -196,8 +231,40 @@ export const AssistantBubble: FC<{
                             </Text>
                         </Stack>
                     ) : (
-                        <AiChartVisualization vizData={vizQuery.data} />
+                        <AiChartVisualization
+                            query={queryExecutionHandle.data.query}
+                            type={queryExecutionHandle.data.type}
+                            vizConfig={message.vizConfigOutput}
+                            results={queryResults}
+                        />
                     )}
+                    <Stack gap="xs">
+                        <ErrorBoundary>
+                            {queryExecutionHandle.data &&
+                                queryExecutionHandle.data.type !==
+                                    AiChartType.CSV && (
+                                    <AgentVisualizationMetricsAndDimensions
+                                        metricQuery={
+                                            queryExecutionHandle.data.query
+                                                .metricQuery
+                                        }
+                                        fieldsMap={
+                                            queryExecutionHandle.data.query
+                                                .fields
+                                        }
+                                    />
+                                )}
+
+                            {message.filtersOutput && (
+                                <AgentVisualizationFilters
+                                    // TODO: fix this using schema
+                                    filters={
+                                        message.filtersOutput as FilterSchemaType
+                                    }
+                                />
+                            )}
+                        </ErrorBoundary>
+                    </Stack>
                 </Paper>
             )}
             <Group gap={0} display={isPreview ? 'none' : 'flex'}>
@@ -253,7 +320,7 @@ export const AssistantBubble: FC<{
                         />
                     </ActionIcon>
                 )}
-                {!!vizQuery.data?.openInExploreUrl && (
+                {!isQueryLoading && !isQueryError && (
                     <Button
                         variant="subtle"
                         color="gray"
@@ -261,7 +328,7 @@ export const AssistantBubble: FC<{
                         aria-label="open in explore"
                         leftSection={<MantineIcon icon={IconExternalLink} />}
                         component={Link}
-                        to={vizQuery.data.openInExploreUrl}
+                        to={openInExploreUrl}
                         target="_blank"
                         style={{
                             color: '#868e96',